requirements.txt
matplotlib==3.7.1
networkx==3.0
numpy==1.24.2
pandas==1.5.3
python-dotenv==1.0.0
scipy==1.11.3
seaborn==0.12.2
